pl_csv: add performer to track info
Just another copy of artist.  This is going to be used in the csv
files since most tools expect the tag to be named performer instead of
artist.

Alternately the rows could be changed with something like
   for row in tlist:
      row['artist'] = row.pop('performer')

This way is (maybe?) a little faster at the expensive of a little
memory.